Do not generate empty store unless forced.
Empty store is a sign of trouble. A corner case is when all the certificates are
distrusted. Make sure a set of trusted certificates is present before generating
the store. This only applies to the system store.

Also, add '--force' (or '-f') option to be able to force the generation of the
store.